+++ CHANGELOG.md
@@ -23,6 +23,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Remove service name from `otelmongodb` configuration and span attributes. (#763)
 
+### Fixed
+
+- Fix deadlocks and race conditions in Sarama instrumentation. (#754)
+
 ## [0.20.0] - 2021-04-23
 
 ### Changed